+++ src/QuickMarcEditor/QuickMarcRecordInfo/QuickMarcRecordInfo.test.js
@@ -5,6 +5,7 @@ import '@folio/stripes-acq-components/test/jest/__mock__';
 
 import { QuickMarcRecordInfo } from './QuickMarcRecordInfo';
 import { RECORD_STATUS_CURRENT } from './constants';
+import { MARC_TYPES } from '../../common/constants';
 
 jest.mock('react-intl', () => ({
   ...jest.requireActual('react-intl'),
@@ -49,5 +50,16 @@ describe('Given Quick Marc Record Info', () => {
 
     expect(getByText('stripes-components.metaSection.source', { exact: false })).toBeDefined();
     expect(getByText('Doe, John', { exact: false })).toBeDefined();
+  });
+
+  describe('when marc type is authority', () => {
+    it('should display corresponding marc title', () => {
+      const { getByText } = renderQuickMarcRecordInfo({
+        correspondingMarcTag: '130',
+        marcType: MARC_TYPES.AUTHORITY,
+      });
+
+      expect(getByText('ui-quick-marc.record.headingType.130')).toBeDefined();
+    });
   });
 });