+++ src/common/constants/api.js
@@ -3,6 +3,7 @@ import { MARC_TYPES } from './marcTypes';
 export const EXTERNAL_INSTANCE_APIS = {
   [MARC_TYPES.BIB]: 'inventory/instances',
   [MARC_TYPES.HOLDINGS]: 'holdings-storage/holdings',
+  [MARC_TYPES.AUTHORITY]: 'authority-storage/authorities',
 };
 
 export const MARC_RECORD_API = 'records-editor/records';