+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * [UIQM-181](https://issues.folio.org/browse/UIQM-181) Add as a default `008` and `852` fields to create MARC Holdings record page.
 * [UIQM-183](https://issues.folio.org/browse/UIQM-183) Add QuickMarcView component.
 * [UIQM-191](https://issues.folio.org/browse/UIQM-191) Use supported `uuid`.
+* [UIQM-142](https://issues.folio.org/browse/UIQM-142) Edit MARC authority record via quickMARC.
 
 ## [4.0.3](https://github.com/folio-org/ui-quick-marc/tree/v4.0.3) (2021-11-09)